计算聘用率和续课率

+++ src/main/java/com/education/hjj/bz/controller/PayController.java
@@ -385,7 +385,7 @@ public class PayController {
 	        		
 	        		teacherPo.setEmployCount(teacherVo.getEmployCount() + 1);
 
-	        		int chooseCount = teacherVo.getChooseCount();
+	        		double chooseCount = teacherVo.getChooseCount();
 
 					double newRate = 0;
 					if (chooseCount != 0) {
